Crate cmtool_data

Modules

case 🔒

descriptors 🔒

flowmap 🔒

rawdata 🔒

states 🔒

transitioner 🔒

Structs

CCMCaseInfo

A type responsible for reading and writing CMCase instances C comparible (binary) files.

CMCase

Represents a case configuration for a computational model analysis.

CMCaseJson

A type responsible for reading and writing CMCase instances to/from JSON files.

DiscontinuousTransitioner

A transitioner that handles discontinuous transitions between flow-map states.

FlowMapDescriptor

FluxFileHeader

Represents the header of a flux file.

HydroState

IterationState

RawDataFlux

Represents a collection of raw flux data along with its header.

RawDataScalar

Represents a collection of raw scalar data along with its header.

RawFlux

Represents a single raw flux interaction.

RawPhase

Represents a phase in a multi-phase process or system.

RawScalar

Represents a single raw scalar value.

ScalarFileHeader

Represents the header of a scalar file.

SimpleTransitioner

Enums

CMAExportType

Legacy type of exported type value in a CM model

CMExportType

Enum-Struct based improvement of ‘CMAExportType’

DataError

Errors that can occur during data operations.

PhaseCM

Phase name in compartment model

TransitionerType

Type of available transitionner

Constants

DEFAULT_CASE_FILE_NAME

Traits

CMCaseReader

A trait for reading a CMCase from a specified path.

CMCaseWriter

A trait for writing a CMCase to a specified path.

FlowMapTransitioner

A trait describing an abstraction over time-indexed flow-map states.

RawData

Functions

get_min_residence_time

Compute the smallest average residence time in compartment

get_transitioner

Create transitioner

linear_index_col_major 🔒

linear_index_row_major 🔒

read_case

Type Aliases

ScalarValueType

Scalar type (float)