cmtool_data

Module rawdata

Structs

FluxFileHeader

Represents the header of a flux file.

RawDataFlux

Represents a collection of raw flux data along with its header.

RawDataScalar

Represents a collection of raw scalar data along with its header.

RawFlux

Represents a single raw flux interaction.

RawPhase

Represents a phase in a multi-phase process or system.

RawScalar

Represents a single raw scalar value.

ScalarFileHeader

Represents the header of a scalar file.

Traits

FromBytes

RawData

ToBytes

Type Aliases

ScalarValueType

Scalar type (float)